refactor(markdown): fold demoteHeadings into markdown_renderer
The previous commit put the shared heading-demotion logic in its own
heading_demotion.ts module, but a dedicated file for ~40 lines is
unnecessary: the logic originally lived in markdown_renderer.ts (as
handleH1), commons is already where it belongs, and trilium-core's HTML
importer already imports from markdown_renderer.js. Export demoteHeadings
from markdown_renderer.ts and drop the separate module; its unit tests move
into markdown_renderer.spec.ts. No behavior change.
